processor/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Image Processing
numpy<2.0.0
Pillow>=10.0.0

# Face Recognition
# face_recognition>=1.3.0
# dlib is a dependency of face_recognition, but often fails to build.

# InsightFace
insightface>=0.7.3
# onnxruntime>=1.16.0
onnxruntime-silicon>=1.16.0; sys_platform == 'darwin' and platform_machine == 'arm64'
onnxruntime>=1.16.0; sys_platform != 'darwin' or platform_machine != 'arm64'

# CLI
tqdm>=4.66.0

# Utilities
python-dotenv>=1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
psutil>=5.9.0

# CORS
starlette==0.27.0
boto3==1.34.0
